PW-3120 keep old logo if new is unavailable

+++ src/ScheduledTask/FetchPaymentMethodLogosHandler.php
@@ -30,6 +30,7 @@ use Adyen\Shopware\Service\ConfigurationService;
 use Exception;
 use Psr\Log\LoggerAwareTrait;
 use Shopware\Core\Checkout\Payment\PaymentMethodEntity;
+use Shopware\Core\Content\Media\File\MediaFile;
 use Shopware\Core\Content\Media\MediaService;
 use Shopware\Core\Framework\Context;
 use Shopware\Core\Framework\DataAbstractionLayer\EntityRepositoryInterface;
@@ -58,19 +59,25 @@ class FetchPaymentMethodLogosHandler extends ScheduledTaskHandler
      * @var EntityRepositoryInterface
      */
     private $mediaRepository;
+    /**
+     * @var bool
+     */
+    private $enableUrlUploadFeature;
 
     public function __construct(
         EntityRepositoryInterface $scheduledTaskRepository,
         ConfigurationService $configurationService,
         MediaService $mediaService,
         EntityRepositoryInterface $paymentMethodRepository,
-        EntityRepositoryInterface $mediaRepository
+        EntityRepositoryInterface $mediaRepository,
+        bool $enableUrlUploadFeature = true
     ) {
         parent::__construct($scheduledTaskRepository);
         $this->configurationService = $configurationService;
         $this->mediaService = $mediaService;
         $this->paymentMethodRepository = $paymentMethodRepository;
         $this->mediaRepository = $mediaRepository;
+        $this->enableUrlUploadFeature = $enableUrlUploadFeature;
     }
 
     public static function getHandledMessages(): iterable
@@ -80,6 +87,11 @@ class FetchPaymentMethodLogosHandler extends ScheduledTaskHandler
 
     public function run(): void
     {
+        if (!$this->enableUrlUploadFeature) {
+            $this->logger->debug('Configuration `shopware.media.enable_url_upload_feature` is disabled.');
+            return;
+        }
+
         $context = Context::createDefaultContext();
         $environment = $this->configurationService->getEnvironment();
         foreach (PaymentMethods::PAYMENT_METHODS as $identifier) {
@@ -97,40 +109,62 @@ class FetchPaymentMethodLogosHandler extends ScheduledTaskHandler
             if ($result->getTotal() === 0) {
                 continue;
             }
-            /** @var PaymentMethodEntity $paymentMethodEntity */
-            $paymentMethodEntity = $result->getEntities()->first();
+
+            // Skip if the remote file is temporarily unavailable.
+            $media = $this->fetchLogoFromUrl($paymentMethod, $environment);
+            if (!$media) {
+                continue;
+            }
 
             // Delete old associated media.
+            /** @var PaymentMethodEntity $paymentMethodEntity */
+            $paymentMethodEntity = $result->getEntities()->first();
             $mediaId = $paymentMethodEntity->getMediaId();
             if ($mediaId) {
                 $this->mediaRepository->delete([['id' => $mediaId]], $context);
             }
 
-            $this->fetchAndAttachLogo($paymentMethod, $paymentMethodEntity->getId(), $context, $environment);
+            $this->attachLogoToPaymentMethod(
+                $media,
+                $context,
+                strtolower($paymentMethod->getGatewayCode()),
+                $paymentMethodEntity->getId()
+            );
         }
     }
 
-    private function fetchAndAttachLogo(
-        PaymentMethodInterface $paymentMethod,
-        string $paymentMethodEntityId,
-        Context $context,
-        string $environment
-    ): void {
+    private function fetchLogoFromUrl(PaymentMethodInterface $paymentMethod, string $environment): ?MediaFile
+    {
         $source = sprintf(
             'https://checkoutshopper-%s.adyen.com/checkoutshopper/images/logos/medium/%s',
             $environment,
             $paymentMethod->getLogo()
         );
-        $request = new Request();
-        $request->query->set('extension', 'png');
-        $request->request->set('url', $source);
-        $request->headers->set('content-type', 'application/json');
+        $media = null;
+        try {
+            $request = new Request();
+            $request->query->set('extension', 'png');
+            $request->request->set('url', $source);
+            $request->headers->set('content-type', 'application/json');
 
-        $media = $this->mediaService->fetchFile($request);
+            $media = $this->mediaService->fetchFile($request);
+        } catch (Exception $exception) {
+            $this->logger->warning(sprintf('The URL %s could not be reached.', $source));
+        }
+
+        return $media;
+    }
+
+    private function attachLogoToPaymentMethod(
+        MediaFile $media,
+        Context $context,
+        string $filename,
+        string $paymentMethodEntityId
+    ): void {
         $mediaId = $this->mediaService->createMediaInFolder('adyen', $context, false);
         $this->mediaService->saveMediaFile(
             $media,
-            strtolower($paymentMethod->getGatewayCode()),
+            $filename,
             $context,
             'adyen',
             $mediaId
